Jenkinsfile: change clone method on merge path
The 'checkout scm' command checks out the target source in a detached
HEAD. The 'git push' command was expecting a checked out branch and thus
failed because of an undefined reference. There were several ways to fix
this but the simpliest was to do a 'git checkout' instead of the
'checkout scm' command.